Allow no S3 object versions.
If versioning is not enabled in the bucket, after put object there will
be no version id. In this case, do not raise error and continue.

When update lambda function, if s3_object_version is None, continue to
invoke update_function_code() without s3 object version. That would be
okay as it is optional.